#130 make start method of CDI module final
Making the default start method in the MvvmfxCdiApplication class final and do the init logic for primary Stage injection there instead of the old `makePrimaryStageInjectable` method.
This way it's not the task of the user anymore to think about injection of the primary stage.

Instead of the default `start` method we have now a new `startMvvmfx` method. This is equal to the mvvmfx-guice module. See issue #136 for a discussion of this change