refactor CompositeValidator and CompositeValidationResult
Due to a change in 1.5.1 it wasn't possible to use subclasses of ValidationMessage together with CompositeValidator anymore.
The reason was an internal refactoring of the CompositeValidationResult class.
In this commit the class was refactored again and now uses another logic to both fix the original bug and still make
usage of custom ValidationMessage types possible.

Fix for #415